+++ spec/integration/resource/api600/c7000/server_profile/create_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::ServerProfile
+klass = OneviewSDK::API600::C7000::ServerProfile
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:storage_system_ip) { $secrets['storage_system1_ip'] }
   let(:server_hardware_type_name) { SERVER_HARDWARE_TYPE_NAME }
-  include_examples 'ServerProfileCreateExample', 'integration api500 context'
+  include_examples 'ServerProfileCreateExample', 'integration api600 context'
 end